There is no point in using an uint32 instead of int32 as th result from recv is an int anyway, so the extra range of the uint32 is not needed.
git-svn-id: svn://svn.code.sf.net/p/themis/code@717 ecaf47ba-bb13-0410-a3b6-899b003a5108

+++ trunk/themis/framework/connection.cpp
@@ -863,7 +863,7 @@ void Connection::RetrieveData() {
 			total_bytes_received+=bytes;
 			TCPMan->AddBytesReceived(bytes);
 			AppendData(ReadBuffer, bytes);
-			if ((uint32)bytes == BufferSize) {
+			if (bytes == BufferSize) {
 				// We might have been able to receive more if we had a bigger buffer
 				// Make it a bit bigger
 				BufferSize += InitialBufferSize;
